pyproject.toml
[tool.poetry]
name = "fooocus-tg"
version = "0.1.0"
description = ""
authors = ["fauzaanu <fauzaangasim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
